fb_receipts/requirements.txt
facebook-business>=19.0.0
pymysql>=1.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
reportlab>=4.0.0
playwright>=1.40.0
